+++ files/reload.sh
@@ -63,7 +63,7 @@ echo "Restoring the saved sets"
 [ -e /etc/sptables/data/whitelist.save ] && $IPSET restore -! -f /etc/sptables/data/whitelist.save
 [ -e /etc/sptables/data/blacklist.save ] && $IPSET restore -! -f /etc/sptables/data/blacklist.save
 [ -e /etc/sptables/data/proxylist.save ] && $IPSET restore -! -f /etc/sptables/data/proxylist.save
-[ -e /etc/sptables/data/crawlerlist.save ] && $IPSET restore -! -f /etc/sptables/data/crawlerlist.save
+[ -e /etc/sptables/data/seclist.save ] && $IPSET restore -! -f /etc/sptables/data/seclist.save
 [ -e /etc/sptables/data/bogonlist.save ] && $IPSET restore -! -f /etc/sptables/data/bogonlist.save
 [ -e /etc/sptables/data/banlist.save ] && $IPSET restore -! -f /etc/sptables/data/banlist.save