+++ files/stop.sh
@@ -81,9 +81,9 @@ if $IPSET list proxylist >/dev/null 2>&1; then
 	$IPSET save proxylist -f /etc/sptables/data/proxylist.save
 	$IPSET destroy proxylist
 fi
-if $IPSET list crawlerlist >/dev/null 2>&1; then
-	$IPSET save crawlerlist -f /etc/sptables/data/crawlerlist.save
-	$IPSET destroy crawlerlist
+if $IPSET list seclist >/dev/null 2>&1; then
+	$IPSET save seclist -f /etc/sptables/data/seclist.save
+	$IPSET destroy seclist
 fi
 if $IPSET list bogonlist >/dev/null 2>&1; then
 	$IPSET save bogonlist -f /etc/sptables/data/bogonlist.save