+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Changed IP set structure
 - Swapped IP set names "blacklist" and "banlist". "whitelist" and "blacklist" are now for user, "banlist" is internally used by Iptables rules.
 - Added timeout support to IP sets named "whitelist" and "blacklist"
-- Added IP set named "crawlerlist"
+- Added IP set named "seclist"
 - Updated examples
 
 ## v2.0.0